Rename bit_vectors module
