+++ INVENTORY.md
@@ -9,6 +9,8 @@
 - Provide more usage examples and documentation.
 - Evaluate additional succinct data structures to include.
 - Investigate alternative dense-select index strategies to replace removed `DArrayIndex`.
+- Audit documentation for broken module links after recent restructuring.
+- Ensure README and examples reference the new `bit_vector` module name.
 
 ## Discovered Issues
 - `katex.html` performs manual string replacements; consider DOM-based manipulation.
