--- src/bit_vectors/bit_vector.rs
@@ -1,4 +1,0 @@
-//! Constants for bit vector operations.
-
-/// The number of bits in a machine word.
-pub const WORD_LEN: usize = std::mem::size_of::<usize>() * 8;
