+++ src/int_vectors/compact_vector.rs
@@ -4,8 +4,8 @@
 use anyhow::{anyhow, Result};
 use num_traits::ToPrimitive;
 
-use crate::bit_vectors::data::BitVectorBuilder;
-use crate::bit_vectors::{BitVector, NoIndex};
+use crate::bit_vector::bit_vector::BitVectorBuilder;
+use crate::bit_vector::{BitVector, NoIndex};
 use crate::int_vectors::prelude::*;
 use crate::utils;
